requirements.txt
Django~=4.0.4
django-filter~=21.1